refactor(cb2-6333): refactor all custom sections into their own folder
