+++ src/app/features/tech-record/components/tech-record-summary/tech-record-summary.component.ts
@@ -26,9 +26,9 @@ import cloneDeep from 'lodash.clonedeep';
 import { Store } from '@ngrx/store';
 import { updateEditingTechRecord } from '@store/technical-records';
 import merge from 'lodash.merge';
-import { WeightsComponent } from '@forms/components/weights/weights.component';
+import { WeightsComponent } from '@forms/custom-sections/weights/weights.component';
 import { TechnicalRecordServiceState } from '@store/technical-records/reducers/technical-record-service.reducer';
-import { DimensionsComponent } from '@forms/components/dimensions/dimensions.component';
+import { DimensionsComponent } from '@forms/custom-sections/dimensions/dimensions.component';
 import { PsvNotes } from '@forms/templates/psv/psv-notes.template';
 import { PsvWeight } from '@forms/templates/psv/psv-weight.template';
 import { HgvWeight } from '@forms/templates/hgv/hgv-weight.template';
