+++ src/app/features/test-records/amend/amend-test-records-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { DefectSelectComponent } from '@forms/components/defect-select/defect-select.component';
-import { DefectComponent } from '@forms/components/defect/defect.component';
+import { DefectComponent } from '@forms/custom-sections/defect/defect.component';
 import { CancelEditTestGuard } from '@guards/cancel-edit-test/cancel-edit-test.guard';
 import { RoleGuard } from '@guards/roles.guard';
 import { Roles } from '@models/roles.enum';
