+++ src/app/features/test-records/components/base-test-record/base-test-record.component.ts
@@ -1,5 +1,5 @@
 import { AfterViewInit, Component, EventEmitter, Input, Output, QueryList, ViewChild, ViewChildren } from '@angular/core';
-import { DefectsComponent } from '@forms/components/defects/defects.component';
+import { DefectsComponent } from '@forms/custom-sections/defects/defects.component';
 import { DynamicFormGroupComponent } from '@forms/components/dynamic-form-group/dynamic-form-group.component';
 import { FormNode } from '@forms/services/dynamic-form.types';
 import { Defect } from '@models/defects/defect.model';
