+++ src/app/forms/custom-sections/defect/defect.component.scss
