+++ src/app/forms/custom-sections/defects/defects.component.spec.ts
@@ -6,14 +6,13 @@ import { DynamicFormService } from '@forms/services/dynamic-form.service';
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { State } from '@store/.';
 import { initialAppState } from '@store/.';
-import { DefectSelectComponent } from '../defect-select/defect-select.component';
+import { DefectSelectComponent } from '../../components/defect-select/defect-select.component';
 import { DefectComponent } from '../defect/defect.component';
 import { DefectsComponent } from './defects.component';
 import { ButtonComponent } from '@shared/components/button/button.component';
 import { TruncatePipe } from '@shared/pipes/truncate/truncate.pipe';
 import { TagComponent } from '@shared/components/tag/tag.component';
 import { RouterTestingModule } from '@angular/router/testing';
-import { CustomFormGroup } from '@forms/services/dynamic-form.types';
 
 describe('DefectsComponent', () => {
   let component: DefectsComponent;
